requirements.txt
# Core
pdfplumber==0.10.3
openpyxl==3.1.2
python-dotenv==1.0.1

# Testing
pytest==8.2.0

# Code formatting (optional)
black==24.4.2
isort==5.13.2